# server.js
require('dotenv').config();
const express = require('express');
const cors = require('cors');
const path = require('path');
const fs = require('fs');
const { initDb, closeDb, query } = require('./src/models/db');
const { validateEnv } = require('./src/config/env');
const { startSubscriptionCron } = require('./src/cron');
const { loggerMiddleware } = require('./src/middleware/loggerMiddleware');
const { compressionMiddleware } = require('./src/middleware/compression');
const { owaspSecurityHeaders } = require('./src/middleware/security');
const logger = require('./src/utils/logger');

// ── Startup Environment Check ────────────────────────────────
validateEnv();

const routes = require('./src/routes');

// ── App Setup ────────────────────────────────────────────────
const app = express();

// ── OWASP Security & CORS Configuration ──────────────────────
app.use(cors({
    origin: true,
    credentials: true,
    methods: ['GET', 'POST', 'PUT', 'PATCH', 'DELETE', 'OPTIONS'],
    allowedHeaders: ['Content-Type', 'Authorization', 'X-Platform-Client', 'X-Request-ID', 'X-Idempotency-Key', 'Accept']
}));

app.use(owaspSecurityHeaders);
app.use(express.json({ limit: '10mb' }));

// ── Attach Performance Compression & Logging Middlewares ──────
app.use(compressionMiddleware);
app.use(loggerMiddleware);
app.use(express.static(path.join(__dirname, 'public')));

const PORT = process.env.PORT || 5000;

// ── OpenAPI 3.1 & Interactive Swagger UI Documentation ──────
const openApiSpecPath = path.join(__dirname, 'docs', 'openapi.json');

app.get(['/firebase-test', '/firebase-test.html'], (req, res) => {
    res.sendFile(path.join(__dirname, 'public', 'firebase-test.html'));
});

app.get('/openapi.json', (req, res) => {
    if (fs.existsSync(openApiSpecPath)) {
        res.setHeader('Content-Type', 'application/json');
        res.sendFile(openApiSpecPath);
    } else {
        res.status(404).json({ error: 'OpenAPI specification file not found' });
    }
});

app.get('/api-docs', (req, res) => {
    const swaggerHtml = `
    <!DOCTYPE html>
    <html lang="en">
    <head>
      <meta charset="UTF-8">
      <title>DigiLocal API Documentation - Swagger UI</title>
      <link rel="stylesheet" href="https://unpkg.com/swagger-ui-dist@5/swagger-ui.css" />
      <style>
        body { margin: 0; padding: 0; background: #FAF9F6; }
        .swagger-ui .topbar { background-color: #0A1428; border-bottom: 3px solid #C5A880; }
        .swagger-ui .topbar a { max-width: 300px; }
      </style>
    </head>
    <body>
      <div id="swagger-ui"></div>
      <script src="https://unpkg.com/swagger-ui-dist@5/swagger-ui-bundle.js"></script>
      <script>
        window.onload = function() {
          SwaggerUIBundle({
            url: "/openapi.json",
            dom_id: '#swagger-ui',
            deepLinking: true,
            presets: [
              SwaggerUIBundle.presets.apis,
              SwaggerUIBundle.SwaggerUIStandalonePreset
            ]
          });
        };
      </script>
    </body>
    </html>
    `;
    res.setHeader('Content-Type', 'text/html');
    res.send(swaggerHtml);
});

// ── Mount Centralized Routes Architecture ─────────────────────
app.use(routes);

// ── Legacy Backward-Compatibility Routes ─────────────────────
app.post('/registerVender', (req, res) => {
    req.url = '/api/vendors/register';
    app._router.handle(req, res);
});

app.get('/venderPanel/:venderId', (req, res) => {
    res.redirect(`/api/vendorPanel/${req.params.venderId}`);
});

// ── QR Code Shop Direct Link ─────────────────────────────────
app.get('/shop/:vendorId', async (req, res) => {
    try {
        const { vendorId } = req.params;
        const result = await query(
            `SELECT vendor_id, society_id, store_name FROM vendors WHERE vendor_id = ?`,
            [vendorId]
        );
        if (result.rows.length === 0) {
            return res.status(404).send('<h2>Shop not found</h2>');
        }
        const vendor = result.rows[0];
        const frontendOrigin = process.env.FRONTEND_URL || 'http://localhost:3000';
        res.redirect(`${frontendOrigin}/${vendor.society_id}/${vendor.vendor_id}`);
    } catch (err) {
        logger.error('QR shop redirect error', { error: err.message, vendorId: req.params.vendorId });
        res.status(500).send('<h2>Server error</h2>');
    }
});

// ── Global Error Handling Middleware ─────────────────────────
app.use((err, req, res, next) => {
    logger.error('Unhandled Application Error', {
        error: err.message,
        stack: err.stack,
        url: req.originalUrl,
        method: req.method,
        requestId: req.id
    });
    res.status(err.status || 500).json({
        error: process.env.NODE_ENV === 'production' ? 'Internal Server Error' : err.message,
        requestId: req.id
    });
});

// ── Database Init & Server Boot ──────────────────────────────
async function bootServer() {
    try {
        await initDb();

        startSubscriptionCron();

        const { initSocket } = require('./src/socket');

        const server = app.listen(PORT, () => {
            console.log(`DigiLocal Server running on PORT ${PORT} | Docs: http://localhost:${PORT}/api-docs`, { port: PORT });
        }).on('error', (err) => {
            if (err.code === 'EADDRINUSE') {
                logger.error(`Port ${PORT} is already in use by another Node process. Run 'npx kill-port ${PORT}' or 'taskkill /F /IM node.exe' to free port ${PORT}.`, { port: PORT });
                process.exit(1);
            } else {
                logger.error('Server boot error', { error: err.message });
                process.exit(1);
            }
        });

        initSocket(server);

        const gracefulShutdown = (signal) => {
            logger.info(`Received ${signal}. Initiating graceful shutdown...`, { signal });
            server.close(async () => {
                logger.info('HTTP server closed.');
                await closeDb();
                logger.info('Database connections closed cleanly. Shutdown complete.');
                process.exit(0);
            });
        };

        process.on('SIGTERM', () => gracefulShutdown('SIGTERM'));
        process.on('SIGINT', () => gracefulShutdown('SIGINT'));
    } catch (err) {
        logger.error('Fatal Database initialization error', { error: err.message });
        process.exit(1);
    }
}

// ── Uncaught Exception Handlers ──────────────────────────────
process.on('unhandledRejection', (reason, promise) => {
    logger.error('Unhandled Rejection at Promise', { reason: reason?.message || reason });
});

process.on('uncaughtException', (err) => {
    logger.error('Uncaught Exception thrown', { error: err.message, stack: err.stack });
});

bootServer();